--- src/components/place-card/place-card-props.ts
@@ -1,11 +1,0 @@
-export type PlaceCardProps = {
-  id: string;
-  title: string;
-  type: 'Room' | 'Apartment' | 'Hotel';
-  price: number;
-  isFavorite: boolean;
-  isPremium: boolean;
-  rating: number;
-  previewImage: string;
-  smallImageSrc: string;
-};
