+++ src/components/place-card/place-card.tsx
@@ -3,12 +3,20 @@ import {Routes} from '../../const.ts';
 import {Offer} from '../../types/offer.ts';
 import React from 'react';
 
-function PlaceCard({id, isPremium, previewImage, price, isFavorite, rating, title, type}: Offer): JSX.Element {
+type PlaceCardProps = {
+  placeCard: Offer;
+  fromNear: boolean;
+}
+
+function PlaceCard({placeCard, fromNear}: PlaceCardProps): JSX.Element {
+  const {id, isPremium, previewImage, price, isFavorite, rating, title, type} = placeCard;
   const bookmarkButtonClassName = `${isFavorite ? 'place-card__bookmark-button--active' : ''} button`;
   const bookmarkStatus = `${isFavorite ? 'In bookmarks' : ''} To bookmarks`;
+  const starsWidth = `${rating * 20}%`;
+  const className = `${fromNear ? 'near-places__card' : 'cities__card'} place-card`;
 
   return (
-    <article className="cities__card place-card">
+    <article className={className}>
       {isPremium &&
         <div className="place-card__mark">
           <span>Premium</span>
@@ -24,18 +32,16 @@ function PlaceCard({id, isPremium, previewImage, price, isFavorite, rating, titl
             <b className="place-card__price-value">&euro;{price}</b>
             <span className="place-card__price-text">&#47;&nbsp;night</span>
           </div>
-          {
-            <button className={`place-card__bookmark-button ${bookmarkButtonClassName}`} type="button">
-              <svg className="place-card__bookmark-icon" width="18" height="19">
-                <use xlinkHref="#icon-bookmark"></use>
-              </svg>
-              <span className="visually-hidden">{bookmarkStatus}</span>
-            </button>
-          }
+          <button className={`place-card__bookmark-button ${bookmarkButtonClassName}`} type="button">
+            <svg className="place-card__bookmark-icon" width="18" height="19">
+              <use xlinkHref="#icon-bookmark"></use>
+            </svg>
+            <span className="visually-hidden">{bookmarkStatus}</span>
+          </button>
         </div>
         <div className="place-card__rating rating">
           <div className="place-card__stars rating__stars">
-            <span style={{ width: `${rating * 20}%` }}></span>
+            <span style={{ width: starsWidth }}></span>
             <span className="visually-hidden">Rating</span>
           </div>
         </div>
